Refactored ValentineProposal into a template with props
