+++ components/SuccessModal.tsx
@@ -3,19 +3,18 @@ import { Dialog, DialogContent, DialogHeader, DialogTitle } from "@/components/u
 interface SuccessModalProps {
   isOpen: boolean
   onClose: () => void
+  message: string
 }
 
-export default function SuccessModal({ isOpen, onClose }: SuccessModalProps) {
+export default function SuccessModal({ isOpen, onClose, message }: SuccessModalProps) {
   return (
     <Dialog open={isOpen} onOpenChange={onClose}>
       <DialogContent className="bg-pink-100 border-2 border-pink-500">
         <DialogHeader>
-          <DialogTitle className="text-2xl font-bold text-pink-600">Yayy! 🎉❤️</DialogTitle>
+          <DialogTitle className="text-2xl font-bold text-pink-600"> Yay!!!</DialogTitle>
         </DialogHeader>
         <div className="text-pink-800">
-          <p>I'm glad you said yes, although if you tried to click no, you'd realize it wasn't really a choice! 😁</p>
-          <p>I've made reservations at Hapa Izakaya at 7:30pm!</p>
-          <p>Meet me at mine at 6:30!</p>
+          <p>{message}</p>
         </div>
       </DialogContent>
     </Dialog>
